CompanyClass Refactor
Refactored the CompanyClass a bit so that there are a bunch of methods
to get information about the company easier.
These include but are not limited to:
* GetOwner/IsOwner
* CanAddFounder/AddFounder
* AddMoney/SpendMoney

As well as the ability to easily change the max amount of founders and
concurrent projects that a company can have.

I also had to quickly fix the statements which required direct access to
the lists to the new mthods.