Updated Skills & Viewers ID/Username conversion
* I have removed the need for 2 dictionaries within the
CommandController class to store IDs and Usernames and replaced it with
a single list and some methods for getting an ID from a username and
vice verse.

* Updated the way in which developers store skills. I have created 2
dictionaries. One for DeveloperSkills and one for LeaderSkills. This is
because they are fundamentally different and might affect a company
differently. This also makes it really easy to add new skills to the
game.

* Added an XpRemaining method to SkillClass in case a viewer wants to
know how close a skill is to levelling up.

* Added the ? symbol to the FindObjectOfType<Canvas> within the
TwitchConnection. This basically means if its not null, then do blah
blah. This will stop the null reference exceptions from happening when
it tries to find the canvas after it is already hidden.