Companies are now saved.
Added the ability for companies to be saved and loaded.
Change the Invite section of the commands to less nested if statements to keep a little more organization.
Check to see if your inviting yourself  and added a cute reply.
Renamed LoadDevelopers and SaveDevelopers to just Load and Save as they were doing more than just Developers. May need to break these apart unless the saves start working on an interval to reduce writes.