Major Update
- Added PasteBinClient
- Added pasteBinAPI key to Settings
Need to add a section to the login form, after login, that requests a API key for PasteBin.
- Added custom QuizClient
- Update ProjectManager
Allows Founders to start a company, and accept applicants.
A list of applicants via a PasteBin link.
Allows Developers to apply to join a company in a specific position.
- Created ProjectApplication class
This handles applications to the project, in much the same way as the CompanyInvite class.
- Renamed Canvas to Login
- Added ProjectManager and LoginCanvas to FindObject
- Added WhisperMessage queue
All messages are sent to either the modQueue or messageQueue.
modQueue is emptied first.
Timers are linked to messages as well, so they'll only activate once the relevant message has been sent.
- Updated client
Added a ton of methods to make it easier to queue messages using the same syntax we've been using before for sending messages.
- Added UI Canvas
UI appears after you've logged in, and currently shows how many messages are queued up.
- Added featureList SO
- Added gearList SO
- Added game features
- Added basic gear
- Updated WhisperMessages class
Organised it using nested classes.