Added UI
- Fixed SkillClass
Apparently it doesn't store the information we sent in through the constructor, so I had to feed it the developer and leadershipSkill bool again when awarding XP.
- Added UI elements
Features and feature information doesn't work exactly as planned, but we can fix that later on.
- Added whisper messages
Added more whisper messages to cover some areas where there is no response.