ToLowers are now string.Equals
I have added a method within the StringExtensions class which allows for
a shorter syntax when comparing strings. The EqualsOrdinalIgnoreCase
method compares 2 strings (or string and developer position) making sure
to not be case sensitive.

This is preferred for a few reasons:
* It makes the intent more obvious, to compare 2 strings while ignoring
the case of both.
* It is so much more performant -
http://www.pvladov.com/2012/09/case-insensitive-string-comparison.html
* Makes modifications easier, only have to change the method in a single
location as opposed to many.